% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pbp.R
\name{extract_all_players}
\alias{extract_all_players}
\title{Extract players on plays from NCAA V1 PBP data}
\usage{
extract_all_players(text)
}
\arguments{
\item{text}{The description of events on a given play from the NCAA V1 PBP data.}
}
\value{
A dataframe of all the players involved in a play + which team they belong to
}
\description{
A helper function to identify which players are involved in plays and which team they belong to
}
