% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pbp.R
\name{parse_pbp}
\alias{parse_pbp}
\title{Parse pbp data into participation data}
\usage{
parse_pbp(pbp_text)
}
\arguments{
\item{pbp_text}{A dataframe of pbp data as returned by `scrape_pbp()`.}
}
\value{
A dataframe of lineups and the times that they were on court for.
}
\description{
Takes in a dataframe of PBP data, returns the lineups and the times those lineups were on-court for.
}
\examples{
parse_pbp(scrape_pbp())
whenever a player is subbed in, they have a 1 in that row and a 0
in the row before. whenever a player is subbed out, they have a 0
in that row and a 1 in the row before
we then cascade the 1s and 0s up and down to create map of who is
in the game at any given time
}
