% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_boxscore.R
\name{scrape_boxscore}
\alias{scrape_boxscore}
\title{Scrape box scores for a given game.}
\usage{
scrape_boxscore(game_id = 6048478)
}
\arguments{
\item{game_id}{A unique identifier for an NCAA basketball game, as obtained from `scrape_*bb_scoreboard()` or ncaa.com.}
}
\value{
A dataframe containing the box score for the specified `game_id`.
}
\description{
Obtain the box scores for a specified game.
}
\examples{
(scrape_boxscore)
}
