% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_scoreboards.R
\name{scrape_mbb_scoreboard}
\alias{scrape_mbb_scoreboard}
\title{scrape NCAA MBB Scoreboard}
\usage{
scrape_mbb_scoreboard(date = Sys.Date())
}
\arguments{
\item{date}{The date to obtain games from, in YYYY-MM-DD format. Defaults to the current date.}
}
\value{
a dataframe of the MBB games taking place on `date`.
}
\description{
Obtains all the MBB games that are taking place on a given date
}
\examples{
scrape_mbb_scoreboard("2023-02-21")
}
