% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_pbp.R
\name{scrape_pbp}
\alias{scrape_pbp}
\title{Scrape play by play data}
\usage{
scrape_pbp(game_id = 6048478)
}
\arguments{
\item{game_id}{A unique identifier for an NCAA basketball game, as obtained from `scrape_*bb_scoreboard()` or ncaa.com.}
}
\value{
a dataframe of plays for a given game.
}
\description{
For a given game, obtain the play by play data from data.ncaa.com.
}
\examples{
scrape_pbp()
}
