% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_scoreboards.R
\name{scrape_wbb_scoreboard}
\alias{scrape_wbb_scoreboard}
\title{scrape NCAA WBB Scoreboard}
\usage{
scrape_wbb_scoreboard(date = Sys.Date())
}
\arguments{
\item{date}{The date to obtain games from, in YYYY-MM-DD format. Defaults to the current date.}
}
\value{
a dataframe of the WBB games taking place on `date`.
}
\description{
Obtains all the WBB games that are taking place on a given date
}
\examples{
scrape_wbb_scoreboard("2023-02-21")
}
